Fixed a bug in computing the integral image. Now our method matches OpenCV and the BRISK/FREAK performance should be much better

+++ src/theia/image/descriptor/freak_descriptor.h
@@ -84,7 +84,7 @@ class FreakDescriptorExtractor : public BinaryDescriptorExtractor {
     std::vector<Eigen::BinaryVectorX>* descriptors);
 
  private:
-  uchar MeanIntensity(const Image<uchar>& image, const Image<uchar>& integral,
+  uchar MeanIntensity(const Image<uchar>& image, const Image<int>& integral,
                       const float kp_x, const float kp_y,
                       const unsigned int scale, const unsigned int rot,
                       const unsigned int point) const;
